Move DigitalEdition objects to model, add redirect function
