+++ app/views/digitalpack/country.scala.html
@@ -1,3 +1,4 @@
+@import model.DigitalEdition
 @(edition: DigitalEdition, subscriptionNumberOfMonths: Int = 1)
 
 @main("Select your country | Digital | Subscriptions | The Guardian", bodyClasses = List("is-wide")) {
