+++ app/views/digitalpack/info.scala.html
@@ -1,4 +1,5 @@
-@(edition: DigitalEdition)
+@import model.DigitalEdition
+@(edition: DigitalEdition, redirect: String)
 
 @import model.Benefits
 
@@ -10,7 +11,7 @@
         </div>
         <div class="pricing-cta__action">
             <a class="button button--large button--primary"
-               href="/digital/country"
+               href="@redirect"
                data-test-id="digital-pack-@edition.id-trial-header"
             >Start your free trial</a>
         </div>
