+++ app/views/index_intl.scala.html
@@ -1,3 +1,4 @@
+@import model.DigitalEdition
 @(edition: DigitalEdition)
 
 @main(s"Subscriptions | ${edition.name} | The Guardian", isInternational=true) {
